LayoutTests/fast/js/exception-linenums-in-html-2-expected.txt

This test exercises the source URL and line number that is embedded in JavaScript exceptions, which is displayed in places like the JavaScript console. It differs from exception-linenums-in-html-1.html in that it only works if Bugzilla Bug 3905 has been fixed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof e.sourceURL is "string"
PASS e.line is 21
PASS typeof e.sourceURL is "string"
PASS e.line is 31
PASS typeof e.sourceURL is "string"
PASS e.line is 41
PASS typeof e.sourceURL is "string"
PASS e.line is 51
PASS typeof e.sourceURL is "string"
PASS e.line is 14
PASS successfullyParsed is true

TEST COMPLETE
